third_party/blink/web_tests/flag-specific/disable-layout-ng/inspector-protocol/css/css-get-platform-fonts-expected.txt

Test css.getPlatformFontsForNode method.

Running test: testFirstLetterPseudoClass
Font #0
    name: <Some-family-name-0>
    glyphCount: 29
    isCustomFont: true
Font #1
    name: <Some-family-name-1>
    glyphCount: 7
    isCustomFont: false
Font #2
    name: <Some-family-name-2>
    glyphCount: 1
    isCustomFont: false

Running test: testSelectElementPlatformFonts
Font #0
    name: <Some-family-name-0>
    glyphCount: 27
    isCustomFont: false
